requirements.txt
asgiref==3.10.0
black==25.9.0
click==8.3.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.2
django-safedelete==1.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-writable-nested==0.7.2
isort==7.0.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
PyJWT==2.10.1
pytokens==0.1.10
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
